requirements.txt
Django==4.2.5
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.0.0
django-ckeditor==6.7.0
whitenoise==6.6.0
django-environ==0.11.2
gunicorn==21.2.0
